% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\docType{data}
\name{AgeGroupEnum}
\alias{AgeGroupEnum}
\title{Age Group Indices}
\format{
An object of class \code{numeric} of length 9.
}
\usage{
AgeGroupEnum
}
\description{
List of age group indices.
Wrapper for \code{enum AGE_GROUPS} (in \emph{constant.h}).
\itemize{
\item \code{_0_9} Age group 0-9
\item \code{_10_19} Age group 10-19
\item \code{_20_29} Age group 20-29
\item \code{_30_39} Age group 30-39
\item \code{_40_49} Age group 40-49
\item \code{_50_59} Age group 50-59
\item \code{_60_69} Age group 60-69
\item \code{_70_79} Age group 70-79
\item \code{_80} Age group >80
}
}
\examples{
# Add the age group constants to your workspace:
ag0_9   <- AgeGroupEnum[['_0_9']]
ag10_19 <- AgeGroupEnum[['_10_19']]
ag20_29 <- AgeGroupEnum[['_20_29']]
ag30_39 <- AgeGroupEnum[['_30_39']]
ag40_49 <- AgeGroupEnum[['_40_49']]
ag50_59 <- AgeGroupEnum[['_50_59']]
ag60_69 <- AgeGroupEnum[['_60_69']]
ag70_79 <- AgeGroupEnum[['_70_79']]
ag80    <- AgeGroupEnum[['_80']]
}
\keyword{datasets}
