% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\docType{data}
\name{NETWORK_CONSTRUCTIONS}
\alias{NETWORK_CONSTRUCTIONS}
\title{Network construction types}
\format{
An object of class \code{numeric} of length 5.
}
\usage{
NETWORK_CONSTRUCTIONS
}
\description{
List of network construction types.
Wrapper for \code{enum NETWORK_CONSTRUCTIONS} (in \emph{constant.h}).
\itemize{
  \item{BESPOKE} Wrapper for C enum
    \code{NETWORK_CONSTRUCTION_BESPOKE}
  \item{HOUSEHOLD} Wrapper for C enum
    \code{NETWORK_CONSTRUCTION_HOUSEHOLD}
  \item{WATTS_STROGATZ} Wrapper for C enum
    \code{NETWORK_CONSTRUCTION_WATTS_STROGATZ}
  \item{RANDOM_DEFAULT} Wrapper for C enum
    \code{NETWORK_CONSTRUCTION_RANDOM_DEFAULT}
  \item{RANDOM_DEFAULT} Wrapper for C enum
    \code{NETWORK_CONSTRUCTION_RANDOM}
}
}
\keyword{datasets}
