% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\docType{data}
\name{SAFE_UPDATE_PARAMS}
\alias{SAFE_UPDATE_PARAMS}
\title{Update-able parameters}
\format{
An object of class \code{character} of length 64.
}
\usage{
SAFE_UPDATE_PARAMS
}
\description{
This is the list of parameter names can be updated whilst the model is
running (in \code{Model$update_running_params}).
}
\seealso{
\code{\link{Model}}

\code{\link{COVID19IBM}}
}
\keyword{datasets}
