% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\docType{data}
\name{VACCINE_STATUS}
\alias{VACCINE_STATUS}
\title{Vaccine status}
\format{
An object of class \code{numeric} of length 7.
}
\usage{
VACCINE_STATUS
}
\description{
List of vaccines statuses.
Wrapper for \code{enum VACCINE_STATUS} (in \emph{constant.h}).
\itemize{
  \item{NO_VACCINE} Wrapper for C enum
    \code{NO_VACCINE}
  \item{VACCINE_NO_PROTECTION} Wrapper for C enum
    \code{VACCINE_NO_PROTECTION}
  \item{VACCINE_PROTECTED_FULLY} Wrapper for C enum
    \code{VACCINE_PROTECTED_FULLY}
  \item{VACCINE_PROTECTED_SYMPTOMS} Wrapper for C enum
    \code{VACCINE_PROTECTED_SYMPTOMS}
  \item{VACCINE_WANED_FULLY} Wrapper for C enum
    \code{VACCINE_WANED_FULLY}
  \item{VACCINE_WANED_SYMPTOMS} Wrapper for C enum
    \code{VACCINE_WANED_SYMPTOMS}
}
}
\keyword{datasets}
