% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\docType{data}
\name{VACCINE_TYPES}
\alias{VACCINE_TYPES}
\title{Vaccine types}
\format{
An object of class \code{numeric} of length 2.
}
\usage{
VACCINE_TYPES
}
\description{
List of vaccine types.
Wrapper for \code{enum VACCINE_TYPES} (in \emph{constant.h}).
\itemize{
  \item{FULL} Wrapper for C enum \code{VACCINE_TYPE_FULL}
  \item{SYMPTOM} Wrapper for C enum \code{VACCINE_TYPE_SYMPTOM}
}
}
\keyword{datasets}
